IPv6 patch, v2
git-svn-id: http://svn.ratbox.org/svnroot/trunk@4904 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ include/s_conf.h
@@ -58,8 +58,8 @@ struct ConfItem
   unsigned int     flags;
   int              clients;  /* Number of *LOCAL* clients using this */
   struct irc_inaddr ipnum;
-  unsigned long    ip;       /* only used for I D lines etc. */
-  unsigned long    ip_mask;
+  unsigned int     ip;       /* only used for I D lines etc. */
+  unsigned int     ip_mask;
   char*            name;     /* IRC name, nick, server name, or original u@h */
   char*            host;     /* host part of user@host */
   char*            passwd;
